refactor(architecture): split comparability.py's profile-fingerprint axis (ADR-061)
Batch 2 of the >1500-line abicheck/ file cleanup: comparability.py
(1524 -> 1240 lines) gains a sibling, comparability_profile.py, holding
check_contracts_comparable's profile-fingerprint axis
(_platform_identity_confirmed, _unexplained_profile_fields,
_profile_mismatch_reason, _check_profile_fingerprint_comparable).

comparability.py calls into it via a dynamic importlib.import_module (the
new module needs ComparabilityMismatch and several private
fingerprint-diagnostic helpers back from comparability.py, so a static
two-way import would create the exact
comparability <-> comparability_profile cycle
scripts/check_ai_readiness.py's import-cycle-growth check rejects -- same
shape as the type_reachability.py split). The comparability_sequences.py/
comparability_language_mode.py names the profile axis needs got a
self-aliased re-export in comparability.py too, since several
tests/test_comparability_gate*.py modules still import them from there
directly -- matching that file's own pre-existing re-export idiom for
comparability_fields.py's names.

New module registered in architecture/modules.yaml alongside its siblings.
Verified: check_architecture.py 0 errors, mypy clean, ruff clean, 652
targeted tests pass.
